backend/requirements.txt
aiosmtplib==2.0.2
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
bcrypt==4.0.1
black==26.3.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.6
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
dnspython==2.7.0
email-validator==2.3.0
environs==14.5.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fastapi-mail==1.4.1
fastar==0.8.0
filelock==3.25.2
flake8==7.3.0
fonttools==4.62.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.18
idna==3.10
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
python-jose
kiwisolver==1.5.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.2.1
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==2.4.3
orjson==3.11.3
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pre_commit==4.5.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.4.1
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.37.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==21.2.0
watchfiles==1.1.0
websockets==15.0.1
python-socketio==5.11.4